Add files via upload
TomoSampling.py quantifies the number of particles within a defined volume of a 3D cryo-tomogram based on a given coordinate list (.txt). The principle of this script is to randomly place a virtual box of defined but also tunable volume within the dimension values of the tomograms and measure the coordinate hits in an unbiased way. This script also allows users to define how many repeat measurements to apply at a time (thus how many box measurements). This script was written by Jingwei Xu from Martin Pilhofer laboratory at ETH Zürich.